requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11
django-allauth==0.39.1
django-bootstrap==0.2.4
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.3
whitenoise==4.1.2
